Fixed selection of home template. Refactored site-selection.
